separate app and lib
