+++ pyproject.toml
@@ -4,7 +4,7 @@ version = "0.1.0"
 description = ""
 authors = ["Your Name <you@example.com>"]
 readme = "README.md"
-packages = [ {include = "study_gmm"} ]
+packages = [ {include = "study_gmm", from = "."} ]
 
 [tool.poetry.dependencies]
 python = ">=3.9,<3.13"
